rename sensor_client api -> data_aggregation
