+++ components/sensor_client/src/data_aggregation.py
